### 2021-08-25 #3657 @whedon
[PRE REVIEW]: serp: An R package for smoothing in ordinal regression
**Submitting author:** @ejikeugba (<a href="http://orcid.org/0000-0003-2572-0023">Ejike Ugba</a>)
**Repository:** <a href="https://github.com/ejikeugba/serp" target ="_blank">https://github.com/ejikeugba/serp</a>
**Version:** v.0.2.0
