### 2021-08-25 #3657 @danielskatz
@ejikeugba - your paper header has at least one problem; see the [sample paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) if needed. Please feel free to make changes to your .md file, then use the command `@whedon generate pdf` to try to make a new PDF. Whedon commands need to be the first entry in a new comment.
