+++ JOSS/paper.bib
@@ -1,3 +1,13 @@
+@book{agresti_analysis_2010,
+    	author={Agresti, A},
+    	title={Analysis of Ordinal Categorical Data},
+    	publisher={{John Wiley and Sons}},
+     	url = {https://doi.org/10.1002/9780470594001},
+    	edition = {2nd},
+    	year={2010}
+}
+
+
 @misc{archer_glmnetcr_2014a,
     title = {glmnetcr: Fit a penalized constrained continuation ratio model for predicting an ordinal response},
     author = {Archer, K J},
@@ -20,19 +30,10 @@
 	Author = {Archer, K J and Hou, J and Zhou, Q and Ferber, K and Layne, J G and Gentry, A E},
 	Title = {Ordinalgmifs: An R package for ordinal regression in highdimensional data settings},
 	Journal = {Cancer Informatics},
+        url = {https://doi.org/10.4137/CIN.S20806},
 	Volume = {13},
 	Year = {2014},
 	Pages = {187--195}
-}
-
-
-@book{agresti_analysis_2010,
-    	author={Agresti, A},
-    	title={Analysis of Ordinal Categorical Data},
-    	publisher={{John Wiley and Sons}},
-     	url = {https://doi.org/10.1002/9780470594001},
-    	edition = {2nd},
-    	year={2010}
 }
 
 
@@ -54,16 +55,6 @@
 }
 
 
-@book{fahrmeir_multivariate_2001,
-    	author={Fahrmeir, L and Tutz, G},
-   	title={Multivariate Statistical Modelling Based on Generalized Linear Models},
-    	publisher={Springer},
-     	address ={New York},
-    	edition    = {2nd},
-    	year={2001}
-}
-
-
 @book{fisher_identification_1966,
     	author={Fisher, F},
     	title={The identification problem in econometrics},
@@ -77,24 +68,6 @@
         author = {Fortmann-Roe, S},
         year = {2012},
         url = {http://scott.fortmann-roe.com/docs/BiasVariance.html}
-}
-
-
-@article{friedman_regularization_2010,
-	Author = {Friedman, J and Hastie, T and Tibshirani, R},
-	Title = {Regularization paths for generalized linear models via coordinate descent},
-	Journal = {Journal of statistical software},
-	Volume = {33},
-	Year = {2010},
-	Pages = {1--22}
-}
-
-
-@Misc{goeman_penalized_2014,
-        title = {penalized: L1 (lasso and fused lasso) and L2 (ridge) penalized estimation in GLMs and in the Cox model},
-        author = {Goeman, J and Meijer, R and Chaturvedi, N},
-        year = {2014},
-        url = {https://cran.r-project.org/package=penalized}
 }
 
 
@@ -117,10 +90,11 @@
 
 
 @book{hastie_2009,
-    	author={Hastie, T and Tibshirani, R and Friedman, J},
+    	author={Hastie, T and Tibshirani, R and Friedman, J H},
    	title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
     	publisher={Springer},
     	address ={New York},
+        url={https://books.google.de/books?id=eBSgoAEACAAJ},
     	edition = {2nd},
     	year={2009}
 }
@@ -130,6 +104,7 @@
 	Author = {Hoerl, A and Kennard, R},
 	Title = {Ridge Regression: Biased Estimation for Nonorthogonal Problems},
 	Journal = {Technometrics},
+        url = {https://doi.org/10.1080/00401706.2000.10485983},
 	Volume = {12},
 	Year = {1970},
 	Pages = {55–67}
@@ -140,6 +115,7 @@
        Title = {Bias in parametric estimation: reduction and useful side-effects},
[…]
