+++ JOSS/paper.bib
@@ -2,7 +2,7 @@
     	author={Agresti, A},
     	title={Analysis of Ordinal Categorical Data},
     	publisher={{John Wiley and Sons}},
-     	url = {https://doi.org/10.1002/9780470594001},
+     	doi = {10.1002/9780470594001},
     	edition = {2nd},
     	year={2010}
 }
@@ -30,7 +30,7 @@
 	Author = {Archer, K J and Hou, J and Zhou, Q and Ferber, K and Layne, J G and Gentry, A E},
 	Title = {Ordinalgmifs: An R package for ordinal regression in highdimensional data settings},
 	Journal = {Cancer Informatics},
-        url = {https://doi.org/10.4137/CIN.S20806},
+        doi = {doi.org/10.4137/CIN.S20806},
 	Volume = {13},
 	Year = {2014},
 	Pages = {187--195}
@@ -94,7 +94,7 @@
    	title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
     	publisher={Springer},
     	address ={New York},
-        url={https://books.google.de/books?id=eBSgoAEACAAJ},
+        udoi = {10.1007/978-0-387-84858-7},
     	edition = {2nd},
     	year={2009}
 }
@@ -104,7 +104,7 @@
 	Author = {Hoerl, A and Kennard, R},
 	Title = {Ridge Regression: Biased Estimation for Nonorthogonal Problems},
 	Journal = {Technometrics},
-        url = {https://doi.org/10.1080/00401706.2000.10485983},
+        doi = {10.1080/00401706.2000.10485983},
 	Volume = {12},
 	Year = {1970},
 	Pages = {55–67}
@@ -115,7 +115,7 @@
        Title = {Bias in parametric estimation: reduction and useful side-effects},
        author = {Kosmidis, I},
        Journal = {Wiley Interdisciplinary Reviews: Computational Statistics},
-       url = {https://doi.org/10.1002/wics.1296},
+       doi = {10.1002/wics.1296},
        Volume = {6},
        Year = {2014a},
        Pages = {185-196}
@@ -136,7 +136,7 @@
 	Author = {Peterson, B and Harrell, F E},
 	Title = {Partial proportional odds models for ordinal response variables},
 	Journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
-        url = {https://doi.org/10.2307/2347760},
+        doi = {10.2307/2347760},
 	Volume = {39},
 	Year = {1990},
 	Pages = {205–217}
@@ -187,7 +187,7 @@
 	author = {Tutz, G and Gertheiss, J},
 	title = {Regularized regression for categorical data (with discussion and rejoinder)},
 	journal = {Statistical Modelling},
-        url = {https://doi.org/10.1177/1471082X16642560},
+        doi = {10.1177/1471082X16642560},
 	volume = {16},
 	year = {2016},
 	pages = {161--260}
@@ -207,7 +207,7 @@
 	Author = {Ugba, E R and {M{\"o}rlein}, D and Gertheiss, J},
 	Title = {Smoothing in Ordinal Regression: An Application to Sensory Data},
 	Journal = {Stats},
-        url = {https://doi.org/10.3390/stats4030037},
+        doi = {10.3390/stats4030037},
 	Volume = {4},
 	Year = {2021},
 	Pages = {616--633}
@@ -227,7 +227,7 @@
 	Author = {Wood, S N and Pya, N and Saefken, B},
 	Title = {Smoothing parameter and model selection for general smooth models},
 	Journal = {Journal of the American Statistical Association},
-        url = {https://doi.org/10.1080/01621459.2016.1180986},
[…]
