+++ R/serp.method.R
@@ -63,7 +63,7 @@ print.summary.serp <- function(x, ...){
   max.tun <- FALSE
   nL <- object$ylev
   coef <- as.data.frame(object$coefficients)
-  df <- (object$nobs*(object$ylev-1)) - length(coef[,1L])
+  df <- object$rdf
   na.ac <- length(object$na.action)
   cat("\nCoefficients:\n")
   printCoefmat(coef, digits = 4L, signif.stars = TRUE,
@@ -91,6 +91,7 @@ print.summary.serp <- function(x, ...){
 }
 
 
+
 #' Summary method for a serp object.
 #'
 #' Summarizes the results of the fitted model in a dataframe.
@@ -99,45 +100,14 @@ print.summary.serp <- function(x, ...){
 #' @param ... Not used. Additional summary arguments.
 #' @return
 #' an object of class "summary.serp", a list (depending on the type of
-#' \code{slope} used) with the components itemized below. Note that the
-#' 'components from object' are already defined in the main function.
+#' \code{slope} used). All components from object are already defined
+#' in the \code{\link{serp}} function.
 #' \describe{
-#'   \item{call}{the component from object.}
-#'   \item{link}{the component from object.}
-#'   \item{edf}{the component from object.}
-#'   \item{ylev}{the component from object.}
-#'   \item{nobs}{the component from object.}
-#'   \item{gradient}{the component from object.}
-#'   \item{Hessian}{the component from object.}
-#'   \item{fitted.values}{the component from object.}
-#'   \item{slope}{the component from object.}
-#'   \item{Terms}{the component from object.}
-#'   \item{control}{the component from object.}
-#'   \item{reverse}{the component from object.}
-#'   \item{converged}{the component from object.}
-#'   \item{iter}{the component from object.}
-#'   \item{message}{the component from object.}
-#'   \item{misc}{the component from object.}
-#'   \item{model}{the component from object.}
 #'   \item{coefficients}{the matrix of coefficients, standard errors,
 #'         z-values and p-values.}
-#'   \item{logLik}{the component from object.}
-#'   \item{deviance}{the component from object.}
-#'   \item{aic}{the component from object.}
-#'   \item{bic}{the component from object.}
-#'   \item{contrasts}{the component from object.}
 #'   \item{penalty}{list of penalization information obtained with
 #'         \code{slope} set to "penalize".}
 #'   \item{expcoefs}{the exponentiated coefficients.}
-#'   \item{cvMetric}{the component from object.}
-#'   \item{globalEff}{the component from object.}
-#'   \item{lambda}{the component from object.}
-#'   \item{lambdaGrid}{the component from object}
-#'   \item{na.action}{the component from object.}
-#'   \item{nrFold}{the component from object.}
-#'   \item{testError}{the component from object.}
-#'   \item{tuneMethod}{the component from object.}
-#'   \item{value}{the component from object.}
 #' }
 #'
 #' @seealso \code{\link{serp}}
@@ -185,7 +155,6 @@ summary.serp <- function(object, ...){
   class(object) <- "summary.serp"
   object
 }
-
 
 #' Predict method for object of class 'serp'.
 #'
