+++ R/serp.R
@@ -4,17 +4,18 @@
 #' smooth-effect-on-response penalty (SERP) via a modified Newton-Raphson
 #' algorithm. SERP enables the regularization of the parameter space between
 #' the general and the restricted cumulative models, with a resultant shrinkage
-#' of all subject-specific effects to global effects. The minimum deviance
-#' and CV tuning among others, provide the means of arriving at an optimal
-#' model in a situation where a user-supplied tuning value is not available.
-#' The slope argument allows for the selection of a penalized, unparallel,
+#' of all subject-specific effects to global effects. The Akaike information
+#' critrion (\code{aic}), K-fold cross validation (\code{cv}), among other tuning
+#' aproaches, provide the means of arriving at an optimal tuning parameter in a
+#' in a situation where a user-supplied tuning value is not available.
+#' The \code{slope} argument allows for the selection of a penalized, unparallel,
 #' parallel, or partial slope.
 #'
 #' @usage serp(
 #'      formula,
 #'      link = c("logit", "probit","loglog", "cloglog", "cauchit"),
 #'      slope = c("penalize", "parallel", "unparallel", "partial"),
-#'      tuneMethod = c("deviance", "cv", "finite", "user"),
+#'      tuneMethod = c("aic", "cv", "finite", "user"),
 #'      reverse = FALSE,
 #'      lambdaGrid = NULL,
 #'      cvMetric = c("brier", "logloss", "misclass"),
@@ -38,19 +39,19 @@
 #'   \code{parallel} and \code{partial} denoting the unpenalized non-parallel,
 #'   parallel and semi-parallel coefficients respectively.
 #' @param tuneMethod sets the method of choosing an optimal shrinkage
-#'   parameter, including: \code{deviance}, \code{cv}, \code{finite} and
+#'   parameter, including: \code{aic}, \code{cv}, \code{finite} and
 #'   \code{user}. i.e., the lambda value along parameter shrinkage path at
-#'   which the fit's residual deviance or the cross-validated test error is
+#'   which the fit's AIC or the k-fold cross-validated test error is
 #'   minimal. The finite tuning is used to obtain the model along parameter
 #'   shrinkage for which the log-Likelihood exist (is finite). The 'user'
 #'   tuning supports a user-supplied lambda value.
 #' @param reverse false by default, when true the sign of the linear predictor
 #'   is reversed.
-#' @param lambdaGrid optional user-supplied lambda grid for the cv and deviance
-#'   tuning methods, when the discrete \code{gridType} is chosen. Negative
-#'   range of values are not allowed. A short lambda grid could increase
-#'   computation time assuming large number of predictors and cases in the
-#'   model.
+#' @param lambdaGrid optional user-supplied lambda grid for the \code{aic},
+#'   and \code{cv} tuning methods, when the discrete \code{gridType}
+#'   is chosen. Negative range of values are not allowed. A short lambda grid
+#'   could increase computation time assuming large number of predictors and
+#'   cases in the model.
 #' @param cvMetric sets the performance metric for the cv tuning, with the
 #'   brier score used by default.
 #' @param gridType chooses if a discrete or a continuous lambda grid should be
@@ -170,8 +171,12 @@
 #' \code{anova}, \code{errorMetrics}, etc.
 #'
 #' \describe{
-#'   \item{aic}{the akaike information criterion.}
-#'   \item{bic}{the bayesian information criterion.}
+#'   \item{aic}{the akaike information criterion, with effective degrees of
+#'         freedom obtained from the trace of the generalized hat matrix
[…]
