+++ JOSS/paper.bib
@@ -7,7 +7,6 @@
     	year={2010}
 }
 
-
 @misc{archer_glmnetcr_2014a,
     title = {glmnetcr: Fit a penalized constrained continuation ratio model for predicting an ordinal response},
     author = {Archer, K J},
@@ -16,15 +15,13 @@
     year = {2014}
   }  
   
-  
-  @misc{archer_glmpathcr_2014b,
+@misc{archer_glmpathcr_2014b,
         title = {glmpathcr: Fit a penalized continuation ratio model for predicting an ordinal response},
         author = {Archer, K J},
         url =  {https://cran.r-project.org/package=glmpathcr},
         note = {R package version 1.0.8},
         year = {2014}
   } 
-
 
 @article{archer_ordinalgmifs_2014,
 	Author = {Archer, K J and Hou, J and Zhou, Q and Ferber, K and Layne, J G and Gentry, A E},
@@ -36,7 +33,6 @@
 	Pages = {187--195}
 }
 
-
 @book{bhlmann_statistics_2011,
     	author={Bhlmann, P and Van de Geer, S},
     	title={Statistics for High-Dimensional Data: Methods, Theory and Applications},
@@ -44,7 +40,6 @@
      	address ={Berlin/Heidelberg},
      	year={2011}
 }
-
 
 @misc{Christensen_cumulative_2019,
   	author = {Christensen, R H B},
@@ -54,7 +49,6 @@
   	year = {2019}
 }
 
-
 @book{fisher_identification_1966,
     	author={Fisher, F},
     	title={The identification problem in econometrics},
@@ -62,14 +56,12 @@
      	year={1966}
 }
 
-
 @Misc{Fortmann-Roe_2012,
         title = {Understanding the bias-variance tradeoff},
         author = {Fortmann-Roe, S},
         year = {2012},
         url = {http://scott.fortmann-roe.com/docs/BiasVariance.html}
 }
-
 
 @book{greenberg_advanced_1983,
     	author={Greenberg, Edward, and Webster, Charles E. Jr.},
@@ -79,7 +71,6 @@
     	year={1983}
 }
 
-
 @Manual{harrell_regressionl_2021,
         title = {rms: Regression Modeling Strategies},
         author = {{Harrell Jr}, F E},
@@ -87,7 +78,6 @@
         year = {2021},
         url = {https://CRAN.R-project.org/package=rms}
 }
-
 
 @book{hastie_2009,
     	author={Hastie, T and Tibshirani, R and Friedman, J H},
@@ -99,7 +89,6 @@
     	year={2009}
 }
 
-
 @article{hoerl_ridge_1970,
 	Author = {Hoerl, A and Kennard, R},
 	Title = {Ridge Regression: Biased Estimation for Nonorthogonal Problems},
@@ -109,7 +98,6 @@
 	Year = {1970},
 	Pages = {55–67}
 }
-
 
 @article{Kosmidis_bias_2014,
        Title = {Bias in parametric estimation: reduction and useful side-effects},
@@ -121,7 +109,6 @@
        Pages = {185-196}
 }
 
-
 @article{mcCullagh_regression_1980,
 	Author = {McCullagh, P},
 	Title = {Regression models for ordinal data},
@@ -130,7 +117,6 @@
 	Year = {1980},
 	Pages = {109--142}
 }
-
 
 @article{peterson_partial_1990,
 	Author = {Peterson, B and Harrell, F E},
@@ -142,7 +128,6 @@
 	Pages = {205–217}
 }
 
-
 @Manual{R_language_2021,
[…]
