### 2021-08-31 #3657 @Bisaloo
Hi @DominiqueMakowski :wave:, I think you could bring some interesting insights given your work on statistical R packages. Would you be interested in reviewing this package for JOSS?
