+++ R/serp.method.R
@@ -1,14 +1,12 @@
-#' Print method for object of class 'serp'
+#' Print method for a fitted serp object
 #'
 #' Prints out a vector of coefficients of the fitted model with some
 #' additional goodness-of-fit measures.
 #'
-#' @param x An object of class 'serp'.
+#' @param x An object of class \code{serp}.
 #' @param ... additional arguments.
 #' @return No return value
-#' @seealso \code{\link{serp}}
-#' @examples
-#' # See serp() documentation for examples.
+#' @seealso \code{\link{serp}}, \code{\link{print.summary.serp}}
 #'
 #' @export
 #'
@@ -28,7 +26,7 @@ print.serp <- function(x, ...)
 
   if (max.tun){
     cat("\n")
-    cat("* minimum tuning criterion obtained at lambda upper limit\n")
+    cat("* optimal tuning obtained at lambda grid upper limit\n")
   }
   na.ac <- length(object$na.action)
   if (na.ac > 0){
@@ -40,17 +38,14 @@ print.serp <- function(x, ...)
 }
 
 
-#' Print method for object of class 'summary.serp'
+#' Print method for an object of class summary.serp
 #'
-#' Prints the data frame returned by the \code{summary.serp} method.
+#' Prints out the information supplied via \code{summary.serp} method.
 #'
-#' @param x An object of class 'summary.serp'.
+#' @param x An object of class \code{summary.serp}.
 #' @param ... additional arguments.
 #' @return No return value
-#' @seealso \code{\link{serp}}
-#'
-#' @examples
-#' # See serp() documentation for examples.
+#' @seealso \code{\link{serp}}, \code{\link{print.serp}}
 #'
 #' @export
 #'
@@ -80,7 +75,7 @@ print.summary.serp <- function(x, ...){
   if (object$slope == 'penalize') max.tun <- penalty.print(object, max.tun)
   if (max.tun){
     cat("\n")
-    cat("* minimum tuning criterion obtained at lambda upper limit\n")
+    cat("* optimal tuning obtained at lambda grid upper limit\n")
   }
   if (na.ac > 0){
     cat("\n")
@@ -92,27 +87,33 @@ print.summary.serp <- function(x, ...){
 
 
 
-#' Summary method for a serp object.
+#' Summary method for a fitted serp object.
 #'
-#' Summarizes the results of the fitted model in a dataframe.
+#' This function summarizes the result of a fitted serp object in a dataframe.
 #'
-#' @param object An object of class serp.
+#' @param object An object of class \code{serp}.
 #' @param ... Not used. Additional summary arguments.
 #' @return
-#' an object of class "summary.serp", a list (depending on the type of
-#' \code{slope} used). All components from object are already defined
-#' in the \code{\link{serp}} function.
+#' an object of class \code{summary.serp}. A list (depending on the type of
+#' \code{slope} used) of all model components defined in the \code{\link{serp}},
+#' function with additional components listed below.
 #' \describe{
 #'   \item{coefficients}{the matrix of coefficients, standard errors,
 #'         z-values and p-values.}
+#'   \item{null.deviance}{the deviance for the intercept only model.}
+#'   \item{null.logLik}{the log-likelihood for the intercept only model.}
 #'   \item{penalty}{list of penalization information obtained with
 #'         \code{slope} set to "penalize".}
 #'   \item{expcoefs}{the exponentiated coefficients.}
 #' }
-#'
-#' @seealso \code{\link{serp}}
+#' @seealso \code{\link{anova.serp}}, \code{\link{predict.serp}},
[…]
