+++ cran-comments.md
@@ -1,6 +1,6 @@
 ## What has changed in the current version 
 
-## serp 0.2.0.9001
+## serp 0.2.1
 
 * namespace gets an additional function (update)
 * more examples included in the different function documentation
@@ -9,7 +9,8 @@
 * serp output includes residual degrees of freedom (rdf)
 * function to compute the effective degrees of freedom and rdf from the trace of the generalized hat matrix included
 * serp.summary documentation has its value segment edited
-* new changes made in serp test functions
+* changes made in serp test functions
+* changes made in README.Rmd and README.md
 
 
 ## local R CMD check results
@@ -17,10 +18,41 @@
 ### Test environments
 * local x86_64-w64-mingw32 (64-bit), R 4.0.3 (2020-10-10) ;
 
--- R CMD check results ------------------------------------ serp 0.2.0.9001 ----
-Duration: 3m 29.2s
+-- R CMD check results ----------------------------------------- serp 0.2.1 ----
+Duration: 2m 51.6s
 
 0 errors v | 0 warnings v | 0 notes v
 
 R CMD check succeeded
+
+
+## R-hub builder
+
+### Test environments
+- R-hub windows-x86_64-devel (r-devel)
+- R-hub ubuntu-gcc-release (r-release)
+- R-hub fedora-clang-devel (r-devel)
+
+R CMD check results
+0 errors √ | 0 warnings √ | 0 notes √
+
+Status: OK
+
+
+## win-builder
+
+### Test environments
+- using platform: x86_64-w64-mingw32 (64-bit), R 4.0.3 (2020-10-10)
+- using platform: x86_64-w64-mingw32 (64-bit), R Under development (unstable) (2021-07-21 r80649)
+
+R CMD check results
+
+Status: OK
+
+
+## Downstream dependencies
+- There are currently no downstream dependencies for this package
+
+
+
 
