### 2021-09-04 #3657 @DominiqueMakowski
Hey @Bisaloo thanks for the invite :) I quickly checked the package and because it is focused on a particular type of model (which I have no experience with) I don't feel confident in reviewing it (since I mostly dabble in general post-processing). Apologies and don't hesitate next time!
