### 2021-09-09 #3657 @Bisaloo
Thanks for the suggestions! Looking at their GitHub profiles, they all seem to be able to produce an interesting review for this paper indeed.

@wesleyburr, @OwenWard, @mariobecerra, would one of you be available to review this R package for JOSS. So you don't have to scroll, [this is the manucript](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03657/joss.03657/10.21105.joss.03657.pdf) and [this is the package itself](https://github.com/ejikeugba/serp).

Thanks :slightly_smiling_face:
