### 2021-09-09 #3657 @Bisaloo
Hi @wesleyburr :wave:, thanks for your quick answer!

> @Bisaloo - yes, I can review this, so long as you're ok if I leave it until September 20-24. I'm a bit swamped with the start of term and other commitments next week.

Yes, it's fine. The other reviewer, Bernard Silenou, is not available until September 20 either so we can start the review at this moment. I'll still launch the review process but I have noted you will not be able to start right now. I'll ping you as a reminder on September 20.
