### 2021-09-09 #3705 @whedon
[REVIEW]: serp: An R package for smoothing in ordinal regression
**Submitting author:** @ejikeugba (<a href="http://orcid.org/0000-0003-2572-0023">Ejike Ugba</a>)
**Repository:** <a href="https://github.com/ejikeugba/serp" target ="_blank">https://github.com/ejikeugba/serp</a>
**Version:** v.0.2.0
