### 2021-09-20 #3705 @Bisaloo
:wave::wave: @bernardsilenou @wesleyburr, as agreed this is a friendly reminder about your review for the serp R package and the associated paper.
