### 2021-09-27 #3705 @wesleyburr
The codebase is very clean (as it would have to be, to pass CRAN checks). Nothing to edit in the package or documentation itself, I think, although some modifications should be made to the Readme / community guidelines for the GitHub face and the pkgdown, DOIs added to the references in the paper itself. Three issues opened for the paper author to deal with, then I'm ready to sign off.
