### 2021-09-27 ejikeugba/serp#1 @wesleyburr
Readme - dependencies
@ejikeugba - I suggest editing the README.md and add a short section at the top on dependencies. Yes, CRAN handles it, but it's good to warn the user of incoming forced-installs of packages.

openjournals/joss-reviews#3705
