### 2021-09-27 #3705 @danielskatz
@wesleyburr - please manually create issues in the source repository (https://github.com/ejikeugba/serp) - not here, not using the github convert-to-issue "feature" - this feature doesn't work correctly with JOSS, as the issues are created in the wrong place. (The issues you have created using that feature have been automatically closed, as the JOSS workflow doesn't support them correctly).  Also, if you can mention `openjournals/joss-reviews#3705` in those issue you create in https://github.com/ejikeugba/serp, they will appear here and then be tracked from this issue.
