+++ README.Rmd
@@ -31,7 +31,9 @@ knitr::opts_chunk$set(
 
 Smooth Effects on Response Penalty for CLM
 
-A regularization method for the cumulative link models (CLM). The 'serp' function applies the 'smooth-effect-on-response penalty' (SERP) on the estimates of the general CLM, causing all subject-specific effects associated with each variable in the model to shrink towards a unique global effect. Fitting is done using a modified Newton's method. Several standard model performance and descriptive methods are also available. See [Ugba et al., 2021](https://doi.org/10.3390/stats4030037) and [Tutz and Gertheiss, 2016](https://doi.org/10.1177/1471082X16642560) for more details.
+
+## Overview
+The R package `serp` fits cumulative link models (CLMs) with the `smooth-effect-on-response penalty (SERP)`. The `cumulative model` developed by McCullagh (1980) is probably the most frequently used ordinal model in empirical studies. However, the stochastic ordering property of the general form of the model poses a very serious challenge in most empirical applications of the model. For instance, unstable likelihoods with ill-conditioned parameter space are frequently encountered during the iterative process. `serp` implements a unique regularization method for CLMs that provides the means of smoothing the adjacent categories in the model. At extreme shrinkage, SERP causes all subject-specific effects associated with each variable in the model to shrink towards unique global effects. Fitting is done using a modified Newton's method. Several standard model performance and descriptive methods are also available. See [Ugba et al., 2021](https://doi.org/10.3390/stats4030037) and [Tutz and Gertheiss, 2016](https://doi.org/10.1177/1471082X16642560) for further details on the implemented penalty.
 
 
 ## Example
@@ -73,8 +75,17 @@ summary(f4)
 ```
 
 
-## Installation:
-The released version of serp can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
+## Installation and Use:
+
+### Dependencies
+Before installing `serp` please note that 
+- [[R](https://cran.r-project.org/bin/windows/base/)] version 1.1-4 or higher is required
+- few imports are made from the `ordinal` and `stats` R packages, thus, pre-installed.
+
+See the Description file for list of all external packages associated with `serp`
+
+### Install Instructions
+The released version of `serp` can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
 
 ``` r
 install.packages("serp")
@@ -83,10 +94,9 @@ install.packages("serp")
 or the development version from [GitHub](https://github.com/ejikeugba/serp) with:
 
 ``` r
-# install.packages("devtools")
+if (!require("devtools")) install.packages("devtools")
 devtools::install_github("ejikeugba/serp")
 ```
-
 
 ## Loading:
 ```{r, eval = FALSE}
@@ -94,12 +104,24 @@ library(serp)
 ```
 
 
-## References:
-Ugba, E. R., Mörlein, D. and Gertheiss, J. (2021). Smoothing in Ordinal            Regression: An Application to Sensory Data. *Stats*, 4, 616–633.
-     https://doi.org/10.3390/stats4030037
+## Community Guidelines
 
-Tutz, G. and Gertheiss, J. (2016). Regularized Regression
-     for Categorical Data (With Discussion and Rejoinder).
-     *Statistical Modelling*, 16, 161-260.
-     https://doi.org/10.1177/1471082X16642560
-     
+* Contributing:
+  * Pull requests are welcome and will be included at the discretion of the maintainer. Please submit your contribution to [serp] through the list of [Pull Requests]!
+* Reporting issues and/or seeking support:
[…]
