+++ README.Rmd
@@ -29,15 +29,12 @@ knitr::opts_chunk$set(
 <!-- badges: end -->
 
 
-Smooth Effects on Response Penalty for CLM
-
-
 ## Overview
 The R package `serp` fits cumulative link models (CLMs) with the `smooth-effect-on-response penalty (SERP)`. The `cumulative model` developed by McCullagh (1980) is probably the most frequently used ordinal model in empirical studies. However, the stochastic ordering property of the general form of the model poses a very serious challenge in most empirical applications of the model. For instance, unstable likelihoods with ill-conditioned parameter space are frequently encountered during the iterative process. `serp` implements a unique regularization method for CLMs that provides the means of smoothing the adjacent categories in the model. At extreme shrinkage, SERP causes all subject-specific effects associated with each variable in the model to shrink towards unique global effects. Fitting is done using a modified Newton's method. Several standard model performance and descriptive methods are also available. See [Ugba et al., 2021](https://doi.org/10.3390/stats4030037) and [Tutz and Gertheiss, 2016](https://doi.org/10.1177/1471082X16642560) for further details on the implemented penalty.
 
 
 ## Example
-Consider the cumulative logit model of the wine dataset, where the rating of wine bitterness is predicted with the two treatment factors, temperature and contact.
+Consider the cumulative logit model of the [wine dataset](https://ejikeugba.github.io/serp/reference/wine.html), where the rating of wine bitterness is predicted with the two treatment factors, temperature and contact.
 
 ```R
 ## The unpenalized non-proportional odds model returns unbounded estimates, hence,
@@ -49,8 +46,8 @@ coef(f1)
 
 ```R
 ## The penalized non-proportional odds model with a user-supplied lambda gives 
-## a fully identified model with bounded estimates. A suitable tuning criterion
-## could as well be used to select lambda (e.g., cv) 
+## a fully identified model having bounded estimates. A suitable tuning criterion
+## could as well be used to select lambda (e.g., aic or cv) 
 f2 <- serp(rating ~ temp + contact, slope = "penalize",
            link = "logit", reverse = TRUE, tuneMethod = "user",
            lambda = 1e1 ,data = wine)
@@ -67,22 +64,17 @@ coef(f3)
 ```
 
 ```R
-## The unpenalized proportional odds model with constrained estimates. Using a 
-## very strong lambda in f2 will result in estimates equal to estimates in f4.
+## The unpenalized proportional odds model with constrained estimates. 
+## Under estreme shrinkage, estimates in f2 equal those in this model.  
 f4 <-  serp(rating ~ temp + contact, slope = "parallel",
             reverse = FALSE, link = "logit", data = wine)
 summary(f4)
 ```
 
-
-## Installation and Use:
+## Installation and Use
 
 ### Dependencies
-Before installing `serp` please note that 
-- [[R](https://cran.r-project.org/bin/windows/base/)] version 1.1-4 or higher is required
-- few imports are made from the `ordinal` and `stats` R packages, thus, pre-installed.
-
-See the Description file for list of all external packages associated with `serp`
+Before installing `serp` please ensure to have [[R](https://cran.r-project.org/bin/windows/base/)] version 1.1-4 or higher installed. Moreover, imports from the R `ordinal` and `stats` packages requires having these packages prior installed. See the Description file for more details on dependencies.
 
 ### Install Instructions
 The released version of `serp` can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
[…]
