+++ README.Rmd
@@ -77,7 +77,7 @@ summary(f4)
 Before installing `serp`, it is encouraged to have a recent version of [[R](https://cran.r-project.org/bin/windows/base/)] installed. Also expect an automatic installation of the `ordinal` and `stats` R packages if not already installed.  
 
 
-### Install Instructions
+### Installation
 The released version of `serp` can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
 
 ``` r
