+++ JOSS/paper.bib
@@ -13,31 +13,32 @@
     url =  {https://cran.r-project.org/package=glmnetcr},
     note = {R package version 1.0.6},
     year = {2014}
-  }  
-  
+  }
+
 @misc{archer_glmpathcr_2014b,
         title = {glmpathcr: Fit a penalized continuation ratio model for predicting an ordinal response},
         author = {Archer, K J},
         url =  {https://cran.r-project.org/package=glmpathcr},
         note = {R package version 1.0.8},
         year = {2014}
-  } 
+  }
 
 @article{archer_ordinalgmifs_2014,
 	Author = {Archer, K J and Hou, J and Zhou, Q and Ferber, K and Layne, J G and Gentry, A E},
 	Title = {Ordinalgmifs: An R package for ordinal regression in highdimensional data settings},
 	Journal = {Cancer Informatics},
-        doi = {doi.org/10.4137/CIN.S20806},
+  doi = {doi.org/10.4137/CIN.S20806},
 	Volume = {13},
 	Year = {2014},
 	Pages = {187--195}
 }
 
 @book{bhlmann_statistics_2011,
-    	author={Bhlmann, P and Van de Geer, S},
+    	author={B{\"{u}}hlmann, P and Van de Geer, S},
     	title={Statistics for High-Dimensional Data: Methods, Theory and Applications},
     	publisher={Springer},
      	address ={Berlin/Heidelberg},
+     	doi = {10.1007/978-3-642-20192-9},
      	year={2011}
 }
 
@@ -45,7 +46,7 @@
   	author = {Christensen, R H B},
   	title = {{Cumulative Link Models for Ordinal Regression with the R Package ordinal}},
   	howpublished = {\url{https://cran.r-project.org/web/packages/ordinal/vignettes/clm_article.pdf}},
-	note = {Accessed: 2021-08-25},
+	  note = {Accessed: 2021-08-25},
   	year = {2019}
 }
 
@@ -63,13 +64,17 @@
         url = {http://scott.fortmann-roe.com/docs/BiasVariance.html}
 }
 
-@book{greenberg_advanced_1983,
-    	author={Greenberg, Edward, and Webster, Charles E. Jr.},
-   	title={Advanced Econometrics: A Bridge to the Literature},
-    	publisher={John Wiley},
-    	address ={New York},
-    	year={1983}
+
+@article{bartels_identification_1985,
+       Title = {Identification in Econometrics},
+       author = {Bartels, R},
+       Journal = {The American Statistician},
+       doi = {10.2307/2682804},
+       Volume = {39},
+       Year = {1985},
+       Pages = {102-104}
 }
+
 
 @Manual{harrell_regressionl_2021,
         title = {rms: Regression Modeling Strategies},
@@ -81,10 +86,10 @@
 
 @book{hastie_2009,
     	author={Hastie, T and Tibshirani, R and Friedman, J H},
-   	title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
+   	  title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
     	publisher={Springer},
     	address ={New York},
-        udoi = {10.1007/978-0-387-84858-7},
+      doi = {10.1007/978-0-387-84858-7},
     	edition = {2nd},
     	year={2009}
 }
@@ -93,7 +98,7 @@
 	Author = {Hoerl, A and Kennard, R},
 	Title = {Ridge Regression: Biased Estimation for Nonorthogonal Problems},
 	Journal = {Technometrics},
-        doi = {10.1080/00401706.2000.10485983},
[…]
