+++ R/wine.R
@@ -38,6 +38,7 @@
 #' @references
 #' Randall, J (1989). The analysis of sensory data by generalized linear
 #'     model. \emph{Biometrical Journal}, 31, 781--793.
+#'     https://doi.org/10.1002/bimj.4710310703
 #'
 #' @examples
 #'
