+++ R/serp.R
@@ -157,6 +157,7 @@
 #' McCullagh, P. (1980). Regression Models for Ordinal Data.
 #'     \emph{Journal of the Royal Statistical Society. Series B
 #'     (Methodological)}, 42, pp. 109-142.
+#'     https://doi.org/10.1111/j.2517-6161.1980.tb01109.x
 #'
 #' Tutz, G. and Gertheiss, J. (2016). Regularized Regression
 #'     for Categorical Data (With Discussion and Rejoinder).
