### 2021-10-02 ejikeugba/serp#1 @ejikeugba
README file has been updated as suggested, with package dependencies also mentioned in the installation section.
