### 2021-10-03 ejikeugba/serp#6 @Bisaloo
Make pkgdown website more discoverable
I didn't notice this package had a pkgdown website until it was mentioned in https://github.com/openjournals/joss-reviews/issues/3705. I think it would be helpful to add here in the usual places where people look:

- in the DESCRIPTION file, alongside the GitHub repository. You can see the example of some popular packages such as [ggplot2](https://cran.r-project.org/web/packages/ggplot2/index.html) or [data.table](https://cran.r-project.org/web/packages/data.table/index.html)
- in the GitHub repository details as described [in this stackoverflow answer](https://stackoverflow.com/a/64942409/4439357)

Additionally, you should add it in your `_pkgdown.yml` file to enable cross-linking from other pkgdown websites, as explained [in the downlit `README`](https://downlit.r-lib.org/#cross-package-links).
