+++ DESCRIPTION
@@ -25,7 +25,7 @@ Description: A regularization method for the cumulative link
     implemented here, see, 'Ugba et al. (2021)' <doi:10.3390/stats4030037> and 
     Tutz and Gertheiss (2016) <doi:10.1177/1471082X16642560>.
 License: GPL-2 | file LICENSE
-URL: https://github.com/ejikeugba/serp
+URL: https://github.com/ejikeugba/serp, https://ejikeugba.github.io/serp
 BugReports: https://github.com/ejikeugba/serp/issues
 Depends: 
     R (>= 3.2.0)
