+++ _pkgdown.yml
@@ -1,3 +1,5 @@
+url: https://ejikeugba.github.io/serp
+
 navbar:
   structure:
     left:
