### 2021-10-05 ejikeugba/serp#6 @ejikeugba
Thanks for pointing it out and the examples too. I have added them as suggested.
