### 2021-10-07 ejikeugba/serp#7 @bernardsilenou
Suggestion related to the manuscript
Here are some suggestions to improve the manuscript.

1. Reference section:
- To order the references in the same order that they are cited in the manuscript.
- Line 119, the url of the DOI has an error: https://doi.org/doi.org/10.4137/CIN.S20806
2. Table 1 and Figure can be made to be more self explanatory, there are abbreviations (, vglm, NPOM, lambda, ..) that are explained in the main text only. Please can you add a footer for the aberrations to table 1? The same also applies to the figure 1, you can give the full names.
3. Should line 64 be A Minimal Example ?, “m” is lower case.
4. Can the section called “Statement of Need”  be renamed “State of the Field” ? Since this section describe how this software compares to other commonly-used packages.
5. You can add another section below the introduction called  “Statement of need” that clearly states what problems the software is designed to solve and who are the target audience. Part of the content is already mentioned in the second paragraph of the summary section.

 openjournals/joss-reviews#3705
