+++ JOSS/paper.md
@@ -24,28 +24,32 @@ bibliography: paper.bib
 
 The use of specialized methods for the analysis of categorical data has been on the rise in recent years [@agresti_analysis_2010]. For instance, scientists frequently use the different forms of the ordinal model to analyze relationships between an ordinal response variable and covariates of interest. In particular, the cumulative link model (CLM) propagated by @mcCullagh_regression_1980 finds a wide range of empirical applications in clinical trials, social surveys, market research, etc. However, in high-dimensional settings with a large number of unknown parameters, e.g., if several potential predictors and response categories are modeled, the so-called identification problem could be somewhat unavoidable [@fisher_identification_1966; @bartels_identification_1985]. Thankfully, regularization techniques (see, e.g., @hastie_2009 and @bhlmann_statistics_2011) among other approaches, offer a remedy in such situations.    
 
-The \texttt{serp} software package [@Ugba_serp_2021], available in the Comprehensive R Archive Network ([CRAN](https://CRAN.R-project.org/package=serp)) [@R_language_2021], implements a unique regularization algorithm that enforces smoothness on the adjacent categories of CLMs. The approach provides flexible modeling of CLMs that uses the smooth-effects-on-response penalty (SERP) discussed in @ugba_smoothing_2021 and @tutz_regularized_2016. For instance, when applied to the general form of the cumulative logit model, also known as the non-proportional odds model (NPOM), one obtains the proportional odds model (POM) under extreme parameter shrinkage. While most commonly implemented regularization penalties for CLMs act on the predictors, a form of horizontal penalization, SERP, however, provides a vertically oriented penalization that focuses on the response categories, making it possible to regularize the parameter space between the general and restricted form of the cumulative link models. Fitting in \texttt{serp} is carried out by a modified Newton's optimization method that facilitates an easy convergence and estimation speed. Moreover, as an open-source software package, \texttt{serp} aims to provide a platform for advanced scientific modeling in empirical research. The core features of \texttt{serp}, as well as details about usage, are provided in the software package [documentation](https://cran.r-project.org/web/packages/serp/serp.pdf).
+The \texttt{serp} software package [@Ugba_serp_2021], available in the Comprehensive R Archive Network ([CRAN](https://CRAN.R-project.org/package=serp)) [@R_language_2021], implements a unique regularization algorithm that enforces smoothness on the adjacent categories of CLMs. The approach provides flexible modeling of CLMs that uses the smooth-effects-on-response penalty (SERP) discussed in @ugba_smoothing_2021 and @tutz_regularized_2016. For instance, when applied to the general form of the cumulative logit model, also known as the non-proportional odds model (NPOM), one obtains the proportional odds model (POM) under extreme parameter shrinkage. Fitting in \texttt{serp} is carried out by a modified Newton's optimization method that facilitates an easy convergence and estimation speed. As an open-source software package, \texttt{serp} aims to provide a platform for advanced scientific modeling in empirical research. The core features of \texttt{serp}, as well as details about usage, are provided in the software package [documentation](https://cran.r-project.org/web/packages/serp/serp.pdf).
 
 
 
 # Statement of Need
 
[…]
