+++ JOSS/paper.bib
@@ -7,61 +7,23 @@
     	year={2010}
 }
 
-@misc{archer_glmnetcr_2014a,
-    title = {glmnetcr: Fit a penalized constrained continuation ratio model for predicting an ordinal response},
-    author = {Archer, K J},
-    url =  {https://cran.r-project.org/package=glmnetcr},
-    note = {R package version 1.0.6},
-    year = {2014}
-  }
 
-@misc{archer_glmpathcr_2014b,
-        title = {glmpathcr: Fit a penalized continuation ratio model for predicting an ordinal response},
-        author = {Archer, K J},
-        url =  {https://cran.r-project.org/package=glmpathcr},
-        note = {R package version 1.0.8},
-        year = {2014}
-  }
-
-@article{archer_ordinalgmifs_2014,
-	Author = {Archer, K J and Hou, J and Zhou, Q and Ferber, K and Layne, J G and Gentry, A E},
-	Title = {Ordinalgmifs: An R package for ordinal regression in highdimensional data settings},
-	Journal = {Cancer Informatics},
-  doi = {doi.org/10.4137/CIN.S20806},
-	Volume = {13},
-	Year = {2014},
-	Pages = {187--195}
+@article{mcCullagh_regression_1980,
+	Author = {McCullagh, P},
+	Title = {Regression models for ordinal data},
+	Journal = {Journal of the Royal Statistical Society. Series B (Methodological)},
+	Volume = {42},
+	Year = {1980},
+	doi = {10.1111/j.2517-6161.1980.tb01109.x},
+	Pages = {109--142}
 }
 
-@book{bhlmann_statistics_2011,
-    	author={B{\"{u}}hlmann, P and Van de Geer, S},
-    	title={Statistics for High-Dimensional Data: Methods, Theory and Applications},
-    	publisher={Springer},
-     	address ={Berlin/Heidelberg},
-     	doi = {10.1007/978-3-642-20192-9},
-     	year={2011}
-}
-
-@misc{Christensen_cumulative_2019,
-  	author = {Christensen, R H B},
-  	title = {{Cumulative Link Models for Ordinal Regression with the R Package ordinal}},
-  	howpublished = {\url{https://cran.r-project.org/web/packages/ordinal/vignettes/clm_article.pdf}},
-	  note = {Accessed: 2021-08-25},
-  	year = {2019}
-}
 
 @book{fisher_identification_1966,
     	author={Fisher, F},
     	title={The identification problem in econometrics},
     	publisher={McGraw-Hill},
      	year={1966}
-}
-
-@Misc{Fortmann-Roe_2012,
-        title = {Understanding the bias-variance tradeoff},
-        author = {Fortmann-Roe, S},
-        year = {2012},
-        url = {http://scott.fortmann-roe.com/docs/BiasVariance.html}
 }
 
 
@@ -76,14 +38,6 @@
 }
 
 
-@Manual{harrell_regressionl_2021,
-        title = {rms: Regression Modeling Strategies},
-        author = {{Harrell Jr}, F E},
-        note = {R package version 6.2-0},
-        year = {2021},
-        url = {https://CRAN.R-project.org/package=rms}
-}
-
 @book{hastie_2009,
     	author={Hastie, T and Tibshirani, R and Friedman, J H},
    	  title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
@@ -94,45 +48,25 @@
     	year={2009}
 }
 
-@article{hoerl_ridge_1970,
-	Author = {Hoerl, A and Kennard, R},
-	Title = {Ridge Regression: Biased Estimation for Nonorthogonal Problems},
-	Journal = {Technometrics},
[…]
