+++ JOSS/paper.md
@@ -42,14 +42,12 @@ A common principle behind the regularization procedures with the highlighted sof
 
 # A Minimal Example
 
-The [wine dataset](https://ejikeugba.github.io/serp/reference/wine.html) from @Randall_1989 available in \texttt{serp} software package represents the outcome of a factorial experiment on factors determining the bitterness of wine. Two treatment factors (temperature and contact) with two levels each are provided. Indeed, temperature (warm or cold) and contact (yes or no) between juice and skins can be controlled when crushing grapes during wine production. Nine judges each assessed wine from two bottles from each of the four treatment conditions, making a total of 72 observations. Rating of wine bitterness ranges from 1 = 'least bitter' to 5 = 'most bitter'. A typical modeling framework in this instance is an ordinal regression where the cumulative logit model is used to predict the scores of wine bitterness using the stated treatment factors (i.e., a simple case where all observations are aggregated over bottles and judges, compare e.g., the examples given in the R \texttt{ordinal} package manual [@Christensen_cumulative_2019]). However, the fitted NPOM (see, Table 1) using the \texttt{VGAM}-vglm function [@yee_vgam_2010], for instance,  is not fully identifiable. As observed, unbounded estimates with large standard errors ($\mbox{SE} > 10^3$) were obtained for covariate temperature with respect to the first and last response category, and also the last threshold (Intercept:4). In the actual sense, the absolute values of the estimates and the estimated standard errors for the unbounded parameters seem to diverge to $\infty$ as the stopping criteria of the iterative fitting procedure used become stricter and the number of iterations allowed increases (see, e.g., @Kosmidis_bias_2014). Other software implementations of the same model without regularization run into a similar problem, forcing users to adopt a different modeling approach which may or may not be appropriate for the data. However, by incorporating SERP into the modeling framework, a fully identified model with bounded estimates could be achieved. The realized coefficient paths for the estimated model via \texttt{serp}, using increasing values of the tuning parameter $[\lambda: 10^{-3}, 10^5]$, are shown in \autoref{fig:serpfig}. As observed, with extreme shrinkage all subject-specific estimates (including the unbounded estimates) get smoothed off to parallel estimates (see the dashed blue line in each panel). Thus, estimates between the parameter space of NPOM and POM inclusive could be determined via an appropriate tuning procedure. In this instance, a 5-fold cross-validated tuning with SERP resulted in POM (see Table 1) with no boundary estimates. The intercept parameters too, with no penalty on them, were all stabilized by the smoothing penalty on the remaining coefficients. This obviously, demonstrates the amount of flexibility that could be achieved with \texttt{serp} in fitting cumulative models in general. In addition to the ordinal logit model, \texttt{serp} also fits ordinal models with the probit, loglog, cloglog and cauchit link functions. A penalized partial proportional odds model [@peterson_partial_1990] with the different link functions is also possible with \texttt{serp}.
-
[…]
