+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.7.3
 pkgdown: 1.6.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2021-10-10T10:16Z
+last_built: 2021-10-10T10:40Z
 urls:
   reference: https://ejikeugba.github.io/serp/reference
   article: https://ejikeugba.github.io/serp/articles
