+++ JOSS/paper.md
@@ -36,7 +36,7 @@ Although a very useful tool in regression analysis, the development of regulariz
 # State of the Field
 So far, exiting software packages implementing regularization techniques for ordinal models are mainly extensions of methods typically found in the high-dimensional linear or generalized linear model framework. Mostly, penalties in the tradition of lasso [@tibshirani_regression_1996], ridge [@hoerl_ridge_1970] and elastic net [@zou_regularization_2005] are implemented. The R \texttt{glmnetcr} [@archer_glmnetcr_2014a] and \texttt{glmpathcr} [@archer_glmpathcr_2014b] packages, for instance, fit continuation ratio models with the elastic net penalty, while the R \texttt{rms} package [@harrell_regressionl_2021] fits the cumulative logit model with quadratic (ridge regression) penalty. Also, the R \texttt{ordinalgmifs} package provided by @archer_ordinalgmifs_2014 implements a generalized monotone incremental forward stagewise (GMIFS) algorithm for regularized ordinal regression models, with a solution path akin to the L1 norm (lasso) penalty. Furthermore, the R \texttt{ordinalNet} package [@wurm_regularized_2021] fits the elementwise link multinomial-ordinal (ELMO) class of models with the elastic net using a coordinate descent algorithm.
 
-A common principle behind the regularization procedures with the highlighted software packages is to make the model's maximum likelihood estimates biased a bit towards zero or fully shrunk to zero, of course, with the so-called bias-variance trade-off in mind [@Fortmann-Roe_2012]. However, the regularization method implemented in \texttt{serp} makes a paradigm shift from the traditional form of regularization where estimates are shrunk towards zero. \texttt{serp}  provides a unique form of penalization that shrinks the category effects in the non-proportional cumulative link model towards global effects. Under extreme shrinkage, this form of penalization results in the proportional odds model. A somewhat similar form of smoothing for categorical response data is provided in the \texttt{mgcv} R package [@wood_smoothing_2016] but for un-ordered categories and with smoothing done in the linear predictors. By and large, regularization with \texttt{serp} provides a means of reducing model complexity without necessarily chopping off important features in the model. 
+A common principle behind the regularization procedures with the highlighted software packages is to make the model's maximum likelihood estimates biased a bit towards zero or fully shrunk to zero, of course, with the so-called bias-variance trade-off in mind [@Fortmann-Roe_2012]. However, the regularization method implemented in \texttt{serp} makes a paradigm shift from the traditional form of regularization where estimates are shrunk towards zero. \texttt{serp}  provides a unique form of penalization that shrinks the category effects in the non-proportional cumulative link model towards global effects. Under extreme shrinkage, this form of penalization results in the proportional odds model. A somewhat similar form of smoothing for categorical response data is provided in the R \texttt{mgcv} package [@wood_smoothing_2016] but for un-ordered categories and with smoothing done in the linear predictors. By and large, regularization with \texttt{serp} provides a means of reducing model complexity without necessarily chopping off important features in the model. 
 
 
 
