### 2021-10-10 #3705 @ejikeugba
@bernardsilenou, thank you for the helpful suggestions. The issues mentioned have all been addressed.

Perhaps, it would make sense to re-generate the pdf once again, since changes are made in it. Thanks!
