+++ README.md
@@ -89,9 +89,9 @@ summary(f4)
 ### Dependencies
 
 Before installing `serp`, it is encouraged to have a recent version of
-[R](https://cran.r-project.org/bin/windows/base/)\] installed. Also
-expect an automatic installation of some few dependencies (e.g. the and
-R packages) if not already installed.
+[R](https://cran.r-project.org/bin/windows/base/) installed. Also expect
+an automatic installation of some few dependencies (e.g. the ordinal and
+stats R packages) if not already installed.
 
 ### Installation
 
