+++ README.Rmd
@@ -100,7 +100,7 @@ library(serp)
 
 ## Community Guidelines
 
-Pull requests are welcomed and will be included at the discretion of the maintainer. Please submit your contributions to `serp` through the list of `Pull Requests`. To report issues and/or seek support, please file a new ticket in the [issue](https://github.com/ejikeugba/serp/issues) tracker, and expect a feedback ASAP! 
+Pull requests are welcomed! Please submit your contributions to `serp` through the list of `Pull Requests`, following the [contributing guidelines](https://ejikeugba.github.io/serp/CONTRIBUTING.html). To report issues and/or seek support, please file a new ticket in the [issue](https://github.com/ejikeugba/serp/issues) tracker, and expect a feedback ASAP! 
 
 
 ## Code of Conduct
