+++ JOSS/paper.md
@@ -33,15 +33,15 @@ The \texttt{serp} software package [@Ugba_serp_2021], available in the Comprehen
 Although a very useful tool in regression analysis, the development of regularization methods for categorical response models is still in its elementary stage. This explains why so many popular statistical software packages for ordinal regression do not currently implement such methods. The 'vglm' function in the R \texttt{VGAM} package [@yee_vector_1996; @yee_vgam_2010], for instance, fits the cumulative family of the ordered model, with several link functions, but with no provision for regularization. The same is also true for the 'clm' and 'polr' functions in the R \texttt{ordinal} [@christensen_ordinal_2019] and \texttt{MASS} [@venables_Modern_2002] packages, respectively. Other non R functions for ordinal regression, such as the SAS \texttt{CATMOD} procedure [@SAS_Institute_2018] and the SPSS \texttt{PLUM} procedure [@IBM_spss_2021] also do not currently support regularization. \texttt{serp} fills this gap by providing a very unique and convenient means of regularizing estimates in ordinal models. The software package is intended for diverse forms of scientific applications. It is particularly suited for studies involving ordinal categorical outcomes. Users already conversant with existing R packages for ordinal regression will by no means find it difficult to use \texttt{serp}. The package, moreover, comes with several useful functions that support empirical research while being also equipped with standard model performance and descriptive methods.
 
 
+
 # State of the Field
-So far, exiting software packages implementing regularization techniques for ordinal models are mainly extensions of methods typically found in the high-dimensional linear or generalized linear model framework. Mostly, penalties in the tradition of lasso [@tibshirani_regression_1996], ridge [@hoerl_ridge_1970] and elastic net [@zou_regularization_2005] are implemented. The R \texttt{glmnetcr} [@archer_glmnetcr_2014a] and \texttt{glmpathcr} [@archer_glmpathcr_2014b] packages, for instance, fit continuation ratio models with the elastic net penalty, while the R \texttt{rms} package [@harrell_regressionl_2021] fits the cumulative logit model with quadratic (ridge regression) penalty. Also, the R \texttt{ordinalgmifs} package provided by @archer_ordinalgmifs_2014 implements a generalized monotone incremental forward stagewise (GMIFS) algorithm for regularized ordinal regression models, with a solution path akin to the L1 norm (lasso) penalty. Furthermore, the R \texttt{ordinalNet} package [@wurm_regularized_2021] fits the elementwise link multinomial-ordinal (ELMO) class of models with the elastic net using a coordinate descent algorithm.
+As previously noted, although a topic of intensive research in statistics for decades, regularization methods that are specifically designed for categorical data are relatively new [@ugba_smoothing_2021]. So far, exiting software packages implementing such techniques are mainly extensions of methods typically found in the high-dimensional linear or generalized linear model framework. Mostly, penalties in the tradition of lasso [@tibshirani_regression_1996], ridge [@hoerl_ridge_1970] and elastic net [@zou_regularization_2005] are implemented. The R \texttt{glmnetcr} [@archer_glmnetcr_2014a] and \texttt{glmpathcr} [@archer_glmpathcr_2014b] packages, for instance, fit continuation ratio models with the elastic net penalty, while the R \texttt{rms} package [@harrell_regressionl_2021] fits the cumulative logit model with quadratic (ridge regression) penalty. Also, the R \texttt{ordinalgmifs} package provided by @archer_ordinalgmifs_2014 implements a generalized monotone incremental forward stagewise (GMIFS) algorithm for regularized ordinal regression models, with a solution path akin to the L1 norm (lasso) penalty. Furthermore, the R \texttt{ordinalNet} package [@wurm_regularized_2021] fits the elementwise link multinomial-ordinal (ELMO) class of models with the elastic net using a coordinate descent algorithm.
 
[…]
