+++ JOSS/paper.md
@@ -30,7 +30,7 @@ The \texttt{serp} software package [@Ugba_serp_2021], available in the Comprehen
 
 # Statement of Need
 
-Although a very useful tool in regression analysis, the development of regularization methods for categorical response models is still in its elementary stage. This explains why so many popular statistical software packages for ordinal regression do not currently implement such methods. The 'vglm' function in the R \texttt{VGAM} package [@yee_vector_1996; @yee_vgam_2010], for instance, fits the cumulative family of the ordered model, with several link functions, but with no provision for regularization. The same is also true for the 'clm' and 'polr' functions in the R \texttt{ordinal} [@christensen_ordinal_2019] and \texttt{MASS} [@venables_Modern_2002] packages, respectively. Other non R functions for ordinal regression, such as the SAS \texttt{CATMOD} procedure [@SAS_Institute_2018] and the SPSS \texttt{PLUM} procedure [@IBM_spss_2021] also do not currently support regularization. \texttt{serp} fills this gap by providing a very unique and convenient means of regularizing estimates in ordinal models. The software package is intended for diverse forms of scientific applications. It is particularly suited for studies involving ordinal categorical outcomes. Users already conversant with existing R packages for ordinal regression will by no means find it difficult to use \texttt{serp}. The package, moreover, comes with several useful functions that support empirical research while being also equipped with standard model performance and descriptive methods.
+Although a very useful tool in regression analysis, the development of regularization methods for categorical response models is still in its elementary stage. This explains why so many popular statistical software packages for ordinal regression do not currently implement such methods. The 'vglm' function in the R \texttt{VGAM} package [@yee_vector_1996; @yee_vgam_2010], for instance, fits the cumulative family of the ordered model, with several link functions, but with no provision for regularization. The same is also true for the 'clm' and 'polr' functions in the R \texttt{ordinal} [@christensen_ordinal_2019] and \texttt{MASS} [@venables_Modern_2002] packages, respectively. Other non-R functions for ordinal regression, such as the SAS \texttt{CATMOD} procedure [@SAS_Institute_2018] and the SPSS \texttt{PLUM} procedure [@IBM_spss_2021] also do not currently support regularization. \texttt{serp} fills this gap by providing a very unique and convenient means of regularizing estimates in ordinal models. The software package is intended for diverse forms of scientific applications. It is particularly suited for studies involving ordinal categorical outcomes. Users already conversant with existing R packages for ordinal regression will by no means find it difficult to use \texttt{serp}. The package, moreover, comes with several useful functions that support empirical research while being also equipped with standard model performance and descriptive methods.
 
 
 
@@ -42,7 +42,7 @@ A common principle behind the regularization procedures with the highlighted sof
 
 
 # A Minimal Example
[…]
