### 2021-10-13 ejikeugba/serp#8 @Bisaloo
Use message() instead of cat()
I see that you are using `cat()` whenever you want to communicate with the user. It is considered better practice to use `message()` instead so that it is directed to the correct output (`stderr` instead of `stdout`) and it can be silenced by the user via `suppressMessages()`.

See this great stack overflow answer for a longer explanation: https://stackoverflow.com/a/36700294/4439357

This issue is part of https://github.com/openjournals/joss-reviews/issues/3705.
