### 2021-10-13 ejikeugba/serp#9 @Bisaloo
Use regression tests
I see your package has a good code coverage but upon closer inspection of what is tested, most of the tests related to errors. It would be good to have some regression tests, where you compare the numerical output of some functions to what is known to be correct (e.g., from earlier versions). This prevents the introduction of a bug that would return incorrect values in the future.

Snapshot testing in testthat provides a convenient way to address this issue: https://testthat.r-lib.org/reference/expect_snapshot.html

This issue is part of https://github.com/openjournals/joss-reviews/issues/3705.
