### 2021-10-14 ejikeugba/serp#8 @ejikeugba
Sorry, I didn't mention it, the remaining `cat()` calls in `serp.method` and `serp.misc` can't be replaced. They are used to form the summary tables/outputs, and not necessarily message to users. The output can as well be assigned to a variable if desired.
