+++ JOSS/paper.bib
@@ -290,7 +290,7 @@
        Journal = {Wiley Interdisciplinary Reviews: Computational Statistics},
        doi = {10.1002/wics.1296},
        Volume = {6},
-       Year = {2014a},
+       Year = {2014},
        Pages = {185-196}
 }
 
