+++ R/serp.control.R
@@ -95,3 +95,4 @@ serp.control <- function(
        int.lambdaGrid = int.lambdaGrid,
        lambda.upper.limit = lambda.upper.limit)
 }
+
