+++ JOSS/paper.md
@@ -67,7 +67,7 @@ The [wine dataset](https://ejikeugba.github.io/serp/reference/wine.html) from @R
 Table: Estimates and standard errors (in parenthesis) of regression coefficients of the non-proportional odds model (NPOM) of the wine dataset, having temperature:warm (TW) and contact:yes (CY) as predictors, with estimates obtained using the R \texttt{serp} and \texttt{vglm} functions respectively penalized and non-penalized. 
 
 
-As observed, with extreme shrinkage all subject-specific estimates (including the unbounded estimates) get smoothed off to parallel estimates (see the dashed blue line in each panel). Thus, estimates between the parameter space of NPOM and POM inclusive could be determined via an appropriate tuning procedure. In this instance, a 5-fold cross-validated tuning with SERP resulted in POM (see Table 1) with no boundary estimates. The intercept parameters too, with no penalty on them, were all stabilized by the smoothing penalty on the remaining coefficients. This obviously, demonstrates the amount of flexibility that could be achieved with \texttt{serp} in fitting cumulative models in general. In addition to the ordinal logit model, \texttt{serp} also fits ordinal models with the probit, loglog, cloglog and cauchit link functions. A penalized partial proportional odds model [@peterson_partial_1990] with the different link functions is also possible with \texttt{serp}.
+As observed, with extreme shrinkage all subject-specific estimates (including the unbounded estimates) get smoothed out to parallel estimates (see the dashed blue line in each panel). Thus, estimates between the parameter space of NPOM and POM inclusive could be determined via an appropriate tuning procedure. In this instance, a 5-fold cross-validated tuning with SERP resulted in POM (see Table 1) with no boundary estimates. The intercept parameters too, with no penalty on them, were all stabilized by the smoothing penalty on the remaining coefficients. This obviously, demonstrates the amount of flexibility that could be achieved with \texttt{serp} in fitting cumulative models in general. In addition to the ordinal logit model, \texttt{serp} also fits ordinal models with the probit, loglog, cloglog and cauchit link functions. A penalized partial proportional odds model [@peterson_partial_1990] with the different link functions is also possible with \texttt{serp}.
 
 
 
