### 2021-10-25 4e49678
Use md syntax instead of latex to style pkg name
