### 2021-10-25 e84dae7
Do not add double parenthesis for citations
