+++ JOSS/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary 
 
-The use of specialized methods for the analysis of categorical data has been on the rise in recent years [@agresti_analysis_2010]. For instance, scientists frequently use the different forms of the ordinal model to analyze relationships between an ordinal response variable and covariates of interest. In particular, the cumulative link model (CLM) propagated by @mcCullagh_regression_1980 finds a wide range of empirical applications in clinical trials, social surveys, market research, etc. However, in high-dimensional settings with a large number of unknown parameters, e.g., if several potential predictors and response categories are modeled, the so-called identification problem could be somewhat unavoidable [@fisher_identification_1966; @bartels_identification_1985]. Thankfully, regularization techniques (see, e.g., @hastie_2009 and @bhlmann_statistics_2011) among other approaches, offer a remedy in such situations.    
+The use of specialized methods for the analysis of categorical data has been on the rise in recent years [@agresti_analysis_2010]. For instance, scientists frequently use the different forms of the ordinal model to analyze relationships between an ordinal response variable and covariates of interest. In particular, the cumulative link model (CLM) propagated by @mcCullagh_regression_1980 finds a wide range of empirical applications in clinical trials, social surveys, market research, etc. However, in high-dimensional settings with a large number of unknown parameters, e.g., if several potential predictors and response categories are modeled, the so-called identification problem could be somewhat unavoidable [@fisher_identification_1966; @bartels_identification_1985]. Thankfully, regularization techniques [see, e.g., @bhlmann_statistics_2011;@hastie_2009] among other approaches, offer a remedy in such situations.    
 
 The `serp` software package [@Ugba_serp_2021], available in the Comprehensive R Archive Network ([CRAN](https://CRAN.R-project.org/package=serp)) [@R_language_2021], implements a unique regularization algorithm that enforces smoothness on the adjacent categories of CLMs. The approach provides flexible modeling of CLMs that uses the smooth-effects-on-response penalty (SERP) discussed in @ugba_smoothing_2021 and @tutz_regularized_2016. For instance, when applied to the general form of the cumulative logit model, also known as the non-proportional odds model (NPOM), one obtains the proportional odds model (POM) under extreme parameter shrinkage. Fitting in `serp` is carried out by a modified Newton's optimization method that facilitates an easy convergence and estimation speed. As an open-source software package, `serp` aims to provide a platform for advanced scientific modeling in empirical research. The core features of `serp`, as well as details about usage, are provided in the software package [documentation](https://cran.r-project.org/web/packages/serp/serp.pdf).
 
@@ -42,7 +42,7 @@ A common principle behind the regularization procedures with the highlighted sof
 
 
 # A Minimal Example
[…]
