### 2021-10-25 d051889
Use canonical URL for docs
