+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: serp
 Title: Smooth Effects on Response Penalty for 'CLM'
-Version: 0.2.1
+Version: 0.2.2
 Authors@R: 
     person(given = "Ejike R.",
            family = "Ugba",
