+++ JOSS/paper.bib
@@ -1,315 +1,0 @@
-@book{agresti_analysis_2010,
-    	author={Agresti, A},
-    	title={Analysis of Ordinal Categorical Data},
-    	publisher={{John Wiley and Sons}},
-     	doi = {10.1002/9780470594001},
-    	edition = {2nd},
-    	year={2010}
-}
-
-
-@article{mcCullagh_regression_1980,
-	Author = {McCullagh, P},
-	Title = {Regression models for ordinal data},
-	Journal = {Journal of the Royal Statistical Society. Series B (Methodological)},
-	Volume = {42},
-	Year = {1980},
-	doi = {10.1111/j.2517-6161.1980.tb01109.x},
-	Pages = {109--142}
-}
-
-
-@book{fisher_identification_1966,
-    	author={Fisher, F},
-    	title={The identification problem in econometrics},
-    	publisher={McGraw-Hill},
-     	year={1966}
-}
-
-
-@article{bartels_identification_1985,
-       Title = {Identification in Econometrics},
-       author = {Bartels, R},
-       Journal = {The American Statistician},
-       doi = {10.2307/2682804},
-       Volume = {39},
-       Year = {1985},
-       Pages = {102-104}
-}
-
-
-@book{hastie_2009,
-    	author={Hastie, T and Tibshirani, R and Friedman, J H},
-   	  title={The Elements of Statistical Learning: Data Mining, Inference and Prediction},
-    	publisher={Springer},
-    	address ={New York},
-      doi = {10.1007/978-0-387-84858-7},
-    	edition = {2nd},
-    	year={2009}
-}
-
-
-@book{bhlmann_statistics_2011,
-    	author={B{\"{u}}hlmann, P and Van de Geer, S},
-    	title={Statistics for High-Dimensional Data: Methods, Theory and Applications},
-    	publisher={Springer},
-     	address ={Berlin/Heidelberg},
-     	doi = {10.1007/978-3-642-20192-9},
-     	year={2011}
-}
-
-
-@Misc{Ugba_serp_2021,
-        title = {{serp}: Smooth Effects on Response Penalty for CLM},
-        author = {Ugba, E R},
-        note = {R package version 0.2.0},
-        year = {2021},
-        url = {https://CRAN.R-project.org/package=serp}
-}
-
-
-@Manual{R_language_2021,
-        title = {{R}: A language and environment for statistical computing},
-        author = {{R Core Team}},
-        organization = {R Foundation for Statistical Computing},
-	      address = {Vienna, Austria},
-        year = {2021},
-	      url = {https://www.R-project.org/}
-}
-
-
-@article{ugba_smoothing_2021,
-	Author = {Ugba, E R and {M{\"o}rlein}, D and Gertheiss, J},
-	Title = {Smoothing in Ordinal Regression: An Application to Sensory Data},
-	Journal = {Stats},
-  doi = {10.3390/stats4030037},
-	Volume = {4},
-	Year = {2021},
-	Pages = {616--633}
-}
-
-
-@article{tutz_regularized_2016,
-	author = {Tutz, G and Gertheiss, J},
-	title = {Regularized regression for categorical data (with discussion and rejoinder)},
-	journal = {Statistical Modelling},
-  doi = {10.1177/1471082X16642560},
-	volume = {16},
-	year = {2016},
-	pages = {161--260}
-}
-
-
-@article{yee_vector_1996,
-	Author = {Yee, T W and Wild, C J},
-	Title = {Vector generalized additive models},
-	Journal = {Journal of the Royal Statistical Society. Series B (Methodological)},
-	Volume = {58},
-	Year = {1996},
[…]
