+++ JOSS/paper.md
@@ -1,84 +1,0 @@
----
-title: 'serp: An R package for smoothing in ordinal regression'
-tags:
-  - R
-  - regularization
-  - identification problem
-  - cumulative models
-  - categorical data
-  - proportional odds
-  - shrinkage penalty
-authors:
-  - name: Ejike R. Ugba
-    orcid: 0000-0003-2572-0023
-    affiliation: 1
-affiliations:
- - name: Department of Mathematics and Statistics, School of Economics and Social Sciences, Helmut Schmidt University, Hamburg, Germany
-   index: 1
-date: 25 August 2021
-bibliography: paper.bib
----
-
-
-# Summary 
-
-
-The use of specialized methods for the analysis of categorical data has been on the rise in recent years [@agresti_analysis_2010]. For instance, scientists frequently use the different forms of the ordinal model to analyze relationships between an ordinal response variable and covariates of interest. In particular, the cumulative link model (CLM) propagated by @mcCullagh_regression_1980 finds a wide range of empirical applications in clinical trials, social surveys, market research, etc. However, in high-dimensional settings with a large number of unknown parameters, e.g., if several potential predictors and response categories are modeled, the so-called identification problem could be somewhat unavoidable [@fisher_identification_1966; @bartels_identification_1985]. Thankfully, regularization techniques [see, e.g., @bhlmann_statistics_2011;@hastie_2009] among other approaches, offer a remedy in such situations.
-
-The `serp` software package [@Ugba_serp_2021], available in the Comprehensive R Archive Network ([CRAN](https://CRAN.R-project.org/package=serp)) [@R_language_2021], implements a unique regularization algorithm that enforces smoothness on the adjacent categories of CLMs. The approach provides flexible modeling of CLMs that uses the smooth-effects-on-response penalty (SERP) discussed in @ugba_smoothing_2021 and @tutz_regularized_2016. For instance, when applied to the general form of the cumulative logit model, also known as the non-proportional odds model (NPOM), one obtains the proportional odds model (POM) under extreme parameter shrinkage. Fitting in `serp` is carried out by a modified Newton's optimization method that facilitates an easy convergence and estimation speed. As an open-source software package, `serp` aims to provide a platform for advanced scientific modeling in empirical research. The core features of `serp`, as well as details about usage, are provided in the software package [documentation](https://cran.r-project.org/package=serp/serp.pdf).
-
-
-
-# Statement of Need
-
-Although a very useful tool in regression analysis, the development of regularization methods for categorical response models is still in its elementary stage. This explains why so many popular statistical software packages for ordinal regression do not currently implement such methods. The 'vglm' function in the `VGAM` R package [@yee_vector_1996; @yee_vgam_2010], for instance, fits the cumulative family of the ordered model, with several link functions, but with no provision for regularization. The same is also true for the 'clm' and 'polr' functions in the `ordinal` [@christensen_ordinal_2019] and `MASS` [@venables_Modern_2002] R packages, respectively. Other non-R functions for ordinal regression, such as the SAS `CATMOD` procedure [@SAS_Institute_2018] and the SPSS `PLUM` procedure [@IBM_spss_2021] also do not currently support regularization. `serp` fills this gap by providing a very unique and convenient means of regularizing estimates in ordinal models. The software package is intended for diverse forms of scientific applications. It is particularly suited for studies involving ordinal categorical outcomes. Users already conversant with existing R packages for ordinal regression will by no means find it difficult to use `serp`. The package, moreover, comes with several useful functions that support empirical research while being also equipped with standard model performance and descriptive methods.
-
-
-
-# State of the Field
[…]
