+++ inst/CITATION.cff
@@ -1,0 +1,31 @@
+citHeader("To cite serp in publications use:")
+
+citEntry(entry = "Article",
+  title        = "Smoothing in ordinal regression: An application to sensory data",
+  author       = c(person(given = c("Ejike", "R."),
+                          family = "Ugba"),
+                   person(given = c("Daniel"),
+                          family = "{M{\"o}rlein}"),
+                   person(given = c("Jan"),
+                          family = "Gertheiss"),
+  journal      = "Stats",
+  year         = "2021",
+  volume       = "4",
+  number       = "3",
+  pages        = "616--633",
+  url          = "https://doi.org/10.3390/stats4030037"
+)
+
+
+
+citEntry(entry = "Article",
+  title        = "{serp}: An {R} package for smoothing in ordinal regression",
+  author       = person(given = c("Ejike", "R."),
+                        family = "Ugba"),
+  journal      = "Journal of Open Source Software",
+  year         = "2021",
+  volume       = "6",
+  number       = "66",
+  pages        = "3705",
+  url          = "https://doi.org/10.21105/joss.03705",
+)
