+++ R/serp.anova.R
@@ -50,20 +50,18 @@ anova.serp <- function (object, ..., test = c("Chisq", "none"))
          call. = FALSE)
   mlist <- list(object, ...)
   mclass <- all(unlist(lapply(mlist, class)) == "serp")
-  if (!mclass) stop("object(s) of class other than serp not supported",
+  if (!mclass) stop("object of class other than serp not supported",
                     call. = FALSE)
   ml <- length(mlist)
   dr <- vapply(mlist, function(r) (r$nobs - length(r$coef)), numeric(1))
   slope.type <- vapply(mlist, function(r) (r$slope), character(1))
   if (any(slope.type=="penalize"))
-    stop("available anova-test doesn't support model(s) with penalized slope",
+    stop("test not available for model with penalized slope",
          call. = FALSE)
   hh <- order(dr, decreasing = TRUE)
   mlist <- mlist[hh]
   nm <- nm[hh]
-  if (any(!vapply(mlist, inherits, logical(1), "serp")))
-    stop("input must be an object(s) of class 'serp'", call. = FALSE)
-  fv <- vapply(mlist, function(r) length(r$fitted.values), numeric(1))
+  fv <- vapply(mlist, function(r) length(as.matrix(r$fitted.values)), numeric(1))
   if (any(fv != fv[1L]))
     stop("models were not all fitted to the same ",
          "size of dataset", call. = FALSE)
@@ -84,6 +82,6 @@ anova.serp <- function (object, ..., test = c("Chisq", "none"))
                   "Test", "LR.stat", "df", "Pr(Chi)")
   if (test == "none") res <- res[, -9L]
   class(res) <- c("Anova", "data.frame")
-  attr(res, "heading") <- "Likelihood ratio tests of ordinal models:\n"
+  attr(res, "heading") <- crayon::green("Likelihood ratio tests of ordinal models:\n")
   res
 }
