+++ R/serp.fit.R
@@ -12,13 +12,13 @@ serpfit <- function(x, y, wt, yMtx, link, slope, reverse, control,
       if(class(globalEff) != "formula")
         stop("no object of class formula used in globalEff", call. = FALSE)
       if (grepl('["]', c(globalEff)))
-        stop("variable name(s) in quotes not allowed in globalEff",
+        stop("variable name in quotes not allowed in globalEff",
              call. = FALSE)
       globalEff <- as.character(all.vars(globalEff))
       if (length(globalEff) == 0L)
-        stop("wrong input(s) in 'globalEff'", call. = FALSE)
+        stop("wrong input in 'globalEff'", call. = FALSE)
       if (!length(union(colnames(m), globalEff)) == length(colnames(m)))
-        stop("unknown variable(s) in globalEff", call. = FALSE)
+        stop("unknown variable in globalEff", call. = FALSE)
       if (ncol(x) <= 2L || (ncol(x)-1) == length(globalEff))
         slope <- "parallel"
       if (all.vars(Terms)[[1L]] %in% globalEff)
