+++ NAMESPACE
@@ -1,20 +1,20 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(AIC,serp)
 S3method(BIC,serp)
 S3method(anova,serp)
 S3method(coef,serp)
-S3method(confint,serp)
-S3method(logLik,serp)
-S3method(predict,serp)
 S3method(print,errorMetrics)
 S3method(print,serp)
 S3method(print,summary.serp)
 S3method(summary,serp)
-S3method(vcov,serp)
+export(AIC.serp)
+export(confint.serp)
 export(errorMetrics)
+export(logLik.serp)
+export(predict.serp)
 export(serp)
 export(serp.control)
+export(vcov.serp)
 importFrom(ordinal,dgumbel)
 importFrom(ordinal,pgumbel)
 importFrom(ordinal,qgumbel)
