+++ R/wine.R
@@ -13,26 +13,24 @@
 #' @docType data
 #' @usage wine
 #' @format A data frame with 72 rows and 6 variables:
-#'  \describe{
-#'  \item{\code{response}}{
+#' @return \item{\code{response}}{
 #'    scorings of wine bitterness on a 0---100 continuous scale.
 #'  }
-#'  \item{\code{rating}}{
+#' @return \item{\code{rating}}{
 #'    ordered factor with 5 levels; a grouped version of \code{response}.
 #'  }
-#'  \item{\code{contact}}{
+#' @return \item{\code{contact}}{
 #'    factor with two levels (\code{"no"} and \code{"yes"}).
 #'  }
-#'  \item{\code{temp}}{
+#' @return \item{\code{temp}}{
 #'    temperature: factor with two levels.
 #'  }
-#'  \item{\code{judge}}{
+#' @return \item{\code{judge}}{
 #'    factor with nine levels.
 #'  }
-#'  \item{\code{bottle}}{
+#' @return \item{\code{bottle}}{
 #'    factor with eight levels.
 #'  }
-#'}
 #' @keywords dataset
 #' @source Taken from Randall (1989).
 #' @references
