+++ R/serp.R
@@ -168,60 +168,59 @@
 #'     Regression: An Application to Sensory Data. \emph{Stats}, 4, 616–633.
 #'     https://doi.org/10.3390/stats4030037
 #'
-#' @return An object of class \code{serp} with the components listed below,
+#' @details An object of class \code{serp} with the components listed below,
 #' depending on the type of slope modeled. Other summary methods include:
 #'  \code{summary}, \code{coef}, \code{predict}, \code{vcov},
 #' \code{anova}, \code{errorMetrics}, etc.
 #'
-#' \describe{
-#'   \item{aic}{the akaike information criterion, with effective degrees of
+#' @return \item{aic}{the akaike information criterion, with effective degrees of
 #'         freedom obtained from the trace of the generalized hat matrix
 #'         depending on the tuning parameter.}
-#'   \item{bic}{the bayesian information criterion, with effective degrees of
+#' @return \item{bic}{the bayesian information criterion, with effective degrees of
 #'         freedom obtained from the trace of the generalized hat matrix
 #'         depending on the tuning parameter.}
-#'   \item{call}{the matched call.}
-#'   \item{coef}{a vector of coefficients of the fitted model.}
-#'   \item{converged}{a character vector of fit convergence status.}
-#'   \item{contrasts}{(where relevant) the contrasts used in the model.}
-#'   \item{control}{list of control parameters from \code{serp.control}.}
-#'   \item{cvMetric}{the performance metric used for cv tuning.}
-#'   \item{deviance}{the residual deviance.}
-#'   \item{edf}{the (effective) number of degrees of freedom used by the model}
-#'   \item{fitted.values}{the fitted probabilities.}
-#'   \item{globalEff}{variable(s) in model treated as global effect(s)}
-#'   \item{gradient}{a column vector of gradients for the coefficients at the
+#' @return \item{call}{the matched call.}
+#' @return \item{coef}{a vector of coefficients of the fitted model.}
+#' @return \item{converged}{a character vector of fit convergence status.}
+#' @return \item{contrasts}{(where relevant) the contrasts used in the model.}
+#' @return \item{control}{list of control parameters from \code{serp.control}.}
+#' @return \item{cvMetric}{the performance metric used for cv tuning.}
+#' @return \item{deviance}{the residual deviance.}
+#' @return \item{edf}{the (effective) number of degrees of freedom used by the model}
+#' @return \item{fitted.values}{the fitted probabilities.}
+#' @return \item{globalEff}{variable(s) in model treated as global effect(s)}
+#' @return \item{gradient}{a column vector of gradients for the coefficients at the
 #'         model convergence.}
-#'   \item{Hessian}{the hessian matrix for the coefficients at the model
+#' @return \item{Hessian}{the hessian matrix for the coefficients at the model
 #'         convergence.}
-#'   \item{iter}{number of interactions before convergence or non-convergence.}
-#'   \item{lambda}{a user-supplied single numeric value for the \code{user}
+#' @return \item{iter}{number of interactions before convergence or non-convergence.}
+#' @return \item{lambda}{a user-supplied single numeric value for the \code{user}
 #'         tuning tuning method.}
-#'   \item{lambdaGrid}{a numeric vector of lambda values used to determine the
+#' @return \item{lambdaGrid}{a numeric vector of lambda values used to determine the
 #'         optimum tuning parameter.}
-#'   \item{logLik}{the realized log-likelihood at the model convergence.}
-#'   \item{link}{character vector indicating the link function of the fit.}
-#'   \item{message}{character vector stating the type of convergence obtained}
-#'   \item{misc}{a list to hold miscellaneous fit information.}
-#'   \item{model}{model.frame having variables from formula.}
[…]
