### 2021-11-08 bb33183
Update NEWS.md
