+++ cran-comments.md
@@ -1,29 +1,28 @@
 ## What has changed in the current version 
 
-## serp 0.2.1
-
-* namespace gets an additional function (update)
-* more examples included in the different function documentation
-* shrinkage parameter upper limit set to 1e10 in serp.control
-* aic replaces deviance tuning option in serp tuneMethod 
-* serp output includes residual degrees of freedom (rdf)
-* function to compute the effective degrees of freedom and rdf from the trace of the generalized hat matrix included
-* serp.summary documentation has its value segment edited
-* changes made in serp test functions
-* changes made in README.Rmd and README.md
-
+## serp 0.2.3
+* import from crayon, with colored outputs in returned objects 
+* minor changes in serp documentation
+* provide coefficient() as an alias for coef()
+* README.md updated to include community guidelines and contributors code of conduct
+* changes made in serp test for improved test coverage report
+* bug fix in errorMetrics, with model argument also dropped 
+* add print method and class of objects returned by errorMetrics
+* message() replaces cat() where appropriate
+* add citation for serp package
 
 ## local R CMD check results
 
 ### Test environments
-* local x86_64-w64-mingw32 (64-bit), R 4.0.3 (2020-10-10) ;
+* local x86_64-w64-mingw32 (64-bit), 4.1.1 (2021-08-10) ;
 
--- R CMD check results ----------------------------------------- serp 0.2.1 ----
-Duration: 2m 51.6s
+-- R CMD check results ----------------------------------------- serp 0.2.3 ----
+Duration: 3m 25s
 
 0 errors v | 0 warnings v | 0 notes v
 
 R CMD check succeeded
+
 
 
 ## R-hub builder
@@ -43,7 +42,7 @@ Status: OK
 
 ### Test environments
 - using platform: x86_64-w64-mingw32 (64-bit), R 4.0.3 (2020-10-10)
-- using platform: x86_64-w64-mingw32 (64-bit), R Under development (unstable) (2021-07-21 r80649)
+- using platform: x86_64-w64-mingw32 (64-bit), R Under development (unstable) (2021-11-06 r81149)
 
 R CMD check results
 
@@ -52,7 +51,3 @@ Status: OK
 
 ## Downstream dependencies
 - There are currently no downstream dependencies for this package
-
-
-
-
