+++ README.Rmd
@@ -72,12 +72,7 @@ summary(f4)
 
 #### Installation and Use
 
-##### Dependencies
-Before installing `serp`, it is encouraged to have a recent version of [R](https://cran.r-project.org/bin/windows/base/) installed. Also expect an automatic installation of some few dependencies (e.g. the `ordinal` and `stats` R packages) if not already installed.  
-
-
-##### Installation
-The released version of `serp` can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
+Before installing `serp`, it is encouraged to have a recent version of [R](https://cran.r-project.org/bin/windows/base/) installed. The released version of `serp` can be installed from [CRAN](https://cran.r-project.org/package=serp) with:
 
 ``` r
 install.packages("serp")
@@ -90,7 +85,6 @@ if (!require("devtools")) install.packages("devtools")
 devtools::install_github("ejikeugba/serp")
 ```
 
-##### Loading
 Load `serp` into R environment with:
 ```{r, eval = FALSE}
 library(serp)
