+++ README.md
@@ -24,7 +24,7 @@ status](https://github.com/ejikeugba/serp/workflows/R-CMD-check/badge.svg)](http
 
 ### Overview
 
-`serp`fits cumulative link models (CLMs) with the
+`serp` fits cumulative link models (CLMs) with the
 `smooth-effect-on-response penalty (SERP)`. The `cumulative model`
 developed by McCullagh (1980) is probably the most frequently used
 ordinal model in empirical studies. However, the stochastic ordering
