+++ DESCRIPTION
@@ -40,4 +40,4 @@ Suggests:
     VGAM (>= 1.1-4)
 Encoding: UTF-8
 LazyData: True
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
