### 2022-01-26 cae2ee7
Update NEWS.md
