+++ appveyor.yml
@@ -34,6 +34,7 @@ on_failure:
 
 on_success:
   - Rscript -e "covr::codecov()"
+  - Rscript -e 'pkgdown::build_site()'
 
 artifacts:
   - path: '*.Rcheck\**\*.log'
