+++ .github/workflows/pkgdown.yaml
@@ -47,4 +47,4 @@ jobs:
         run: |
           git config --local user.email "actions@github.com"
           git config --local user.name "GitHub Actions"
-          Rscript -e 'pkgdown::deploy_to_branch(new_process = TRUE)'
+          Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
