+++ tests/testthat/test-penalized-model.R
@@ -53,28 +53,19 @@ test_that("estreem shrinkage with serp results to the
             #3# checks on anova, confinct and vcov functions
             expect_equal(coef(sp5), coef(sp6), check.attributes=FALSE,
                          tolerance=tol)
-            expect_error(anova.serp(sp5),
-                         "no anova implementation yet for a single 'serp' object")
+            expect_error(anova.serp(sp5))
             expect_false(inherits(try(anova.serp(sp5, sp6)), 'try-error'))
-            expect_error(anova.serp(lm(rnorm(50) ~ runif(50))),
-                         "object of class other than serp not supported")
-            expect_error(anova.serp(sp3, sp4),
-                         "test not available for model with penalized slope")
-            expect_error(anova.serp(sp5,
-                                    update(sp6, subset = c(1:40))),
-                         "models were not all fitted to the same size of dataset")
+            expect_error(anova.serp(lm(rnorm(50) ~ runif(50))))
+            expect_error(anova.serp(sp3, sp4))
+            expect_error(anova.serp(sp5, update(sp6, subset = c(1:40))))
             expect_vector(length(anova.serp(sp5, sp6, test = "none")))
-            expect_error(confint.serp(sp5, sp6), "one object at a time allowed")
+            expect_error(confint.serp(sp5, sp6))
             expect_false(inherits(try(confint.serp(sp5)), 'try-error'))
-            expect_error(confint.serp(sp5, level = 1.2),
-                         "values of 'level' should lie between 0 and 1")
-            expect_message(confint.serp(sp5, parm = 0.1),
-                           "argument 'parm' ignored")
-            expect_error(confint.serp(lm(rnorm(50) ~ runif(50))),
-                         "input must be an object of class 'serp'")
+            expect_error(confint.serp(sp5, level = 1.2))
+            expect_message(confint.serp(sp5, parm = 0.1))
+            expect_error(confint.serp(lm(rnorm(50) ~ runif(50))))
             expect_false(inherits(try(vcov.serp(sp5)), 'try-error'))
-            expect_error(vcov.serp(sp1, sp2),
-                         "one object at a time allowed")
+            expect_error(vcov.serp(sp1, sp2))
             expect_vector(AIC.serp(sp5))
             expect_vector(BIC.serp(sp5))
             expect_vector(logLik.serp(sp5))
@@ -92,16 +83,14 @@ test_that("lambda is a single numeric and non-negative value and
             expect_error(
               serp(rating ~ temp + contact, slope = "penalize",
                    link = "loglog", reverse = TRUE, tuneMethod = "user",
-                   lambda = c(0.3,5), data = wine, subset = subs),
-              "lambda should be a single numeric and non-negative value")
+                   lambda = c(0.3,5), data = wine, subset = subs))
 
             expect_error(
               serp(rating ~ temp + contact, link = "probit",
                    slope = "penalize", reverse = FALSE,
                    tuneMethod = "aic",
                    lambdaGrid = c(-3,4),
-                   data = wine, subset = subs),
-              "lambdaGrid must be a non-negative numeric vector of length > 1")
+                   data = wine, subset = subs))
 
             expect_vector(
               serp(rating ~ temp + contact, link = "cloglog",
@@ -115,13 +104,9 @@ test_that("lambda is a single numeric and non-negative value and
                    slope = "penalize", reverse = FALSE,
                    tuneMethod = "deviance",
                    lambdaGrid = c(0,1), control = list(trace=-1),
-                   data = wine, subset = subs),
-              "maxits, eps, maxpen, minP, maxAdjIter, max.half.iter and
-  relTol should all be numeric and non-negative")
-            expect_error(serp.control(misclass.thresh = 5),
[…]
