+++ README.Rmd
@@ -14,7 +14,8 @@ knitr::opts_chunk$set(
 ```
 
 
-# serp <img src='man/figures/hex_logo.png' align="right" height="105" />
+# serp <img src='man/figures/hex_logo.png' align="right" height="139" />
+
 
 <!-- badges: start -->
 
