### 2022-02-15 1d490ba
Update NEWS.md
