+++ cran-comments.md
@@ -1,53 +1,11 @@
-## What has changed in the current version 
-
-## serp 0.2.3
-* import from crayon, with colored outputs in returned objects 
-* minor changes in serp documentation
-* provide coefficient() as an alias for coef()
-* README.md updated to include community guidelines and contributors code of conduct
-* changes made in serp test for improved test coverage report
-* bug fix in errorMetrics, with model argument also dropped 
-* add print method and class of objects returned by errorMetrics
-* message() replaces cat() where appropriate
-* add citation for serp package
-
-## local R CMD check results
+## Local CMD checks
 
 ### Test environments
-* local x86_64-w64-mingw32 (64-bit), 4.1.1 (2021-08-10) ;
+- local x86_64-w64-mingw32 (64-bit), 4.1.2 (2021-11-01) 
 
--- R CMD check results ----------------------------------------- serp 0.2.3 ----
-Duration: 3m 25s
-
+R CMD check results
 0 errors v | 0 warnings v | 0 notes v
 
 R CMD check succeeded
 
 
-
-## R-hub builder
-
-### Test environments
-- R-hub windows-x86_64-devel (r-devel)
-- R-hub ubuntu-gcc-release (r-release)
-- R-hub fedora-clang-devel (r-devel)
-
-R CMD check results
-0 errors √ | 0 warnings √ | 0 notes √
-
-Status: OK
-
-
-## win-builder
-
-### Test environments
-- using platform: x86_64-w64-mingw32 (64-bit), R 4.0.3 (2020-10-10)
-- using platform: x86_64-w64-mingw32 (64-bit), R Under development (unstable) (2021-11-06 r81149)
-
-R CMD check results
-
-Status: OK
-
-
-## Downstream dependencies
-- There are currently no downstream dependencies for this package
