### 2022-02-15 b9f3e75
Update misc.R
