### 2022-02-15 3d3c075
Update NEWS.md
