### 2022-02-16 6662147
Update NEWS.md
