+++ NEWS.md
@@ -1,4 +1,4 @@
-## serp 0.2.3.9002
+## serp 0.2.4
 - errorMetrics() function is modified and no longer exported, an equivalent function with improved functionality is provided by the erroR() function in the gofcat package
 - fix some dependency induced bugs in the package tests
 - remove file License in the description and associated file docs
